fix: Improve header button visibility and consolidate navigation
This commit addresses critical UI issues related to button visibility
in the header and a redundant navigation bar.

1.  **Header Button Visibility:**
    - In `src/components/Header.tsx`, the default text color for
      outline buttons (Reset, Settings, Logout, Login) was changed
      from `text-white` to `text-gray-200`. This ensures these
      buttons, which have a transparent background by default as
      part of the 'outline' variant, are visible against the dark
      (`bg-gray-900`) header background, especially when the page
      content behind the header might be light.

2.  **Consolidate Navigation / Fix Double Nav Bar:**
    - The functionality and links from the separate `MainNavigation.tsx`
      component were merged into `Header.tsx`.
    - `Header.tsx` now includes the primary navigation links (e.g.,
      Builder, Configs, History, Testing, ISO Management) appropriately
      styled for the dark header background.
    - The `MainNavigation.tsx` file is no longer needed.
    - `LFSBuilder.tsx` (the main page component) has been updated to
      remove its direct rendering of `MainNavigation`, now only
      rendering the enhanced `Header` component. This resolves the
      issue of a "double navigation bar" appearing on the main page.
